Fix critical/high review findings in ALM and ANPHON
Addresses defects found in a code review (each verified against the source):

ALM
- constraint.cpp: FC2FIX/FC3FIX collapsed all fixed force constants into a
  single constraint row because ++icount was outside the per-FC loop
  (setFromTriplets summed duplicates; rhs overwritten). Move ++icount inside
  each loop so every fixed FC gets its own row. (regression since 2025-07-03)
- symmetry.cpp: spg_get_dataset() can return NULL on symmetry-detection
  failure; n_operations was dereferenced before any null check -> segfault.
  Add an explicit null check with a clear error message.
- svd.cpp: new double[m * n] (and m*min_mn, min_mn*n, and the column-major
  copy index) multiplied ints, overflowing for large sensing matrices and
  undersizing the buffer. Cast to size_t before multiplying.

ANPHON
- conductivity.cpp: isotope-scattering accumulation was missing its
  temperature loop, so it used the leftover i==ntemp and wrote one element
  past each gamma_total row while dropping isotope scattering from kappa.
  Wrap the addition in the temperature loop.
- conductivity.cpp: kappa_spec was used in the FPH_RTA 3-phonon-only call
  while still null (allocated only afterwards), causing a null deref with
  FPH_RTA>0 and KAPPA_SPEC=1. Allocate kappa_spec before the FPH_RTA block
  and drop the now-duplicate allocation.
- mpi_common.cpp: MPI_Bcast_string copied an arbitrary-length string into a
  fixed char[512] via strcpy (stack buffer overflow). Use a heap buffer
  sized to the broadcast length.
- mpi_common.cpp: MPI_Bcast_SpinClass broadcast a vector<vector<double>>
  via magmom.data(), reinterpreting inner-vector control blocks as doubles.
  Pack into a flat buffer, broadcast, then unpack.

All five edited files compile cleanly (no errors/warnings).

+++ alm/svd.cpp
@@ -38,16 +38,17 @@ auto compute_svd_thin(const Eigen::MatrixXd &A, const bool use_eigen)
         return {U, S, VT};
     } else {
         // LAPACK version: use 'S' for thin U and 'S' for thin VT
-        // Prepare a column-major buffer for Fortran
-        double *A_lap = new double[m * n];
+        // Prepare a column-major buffer for Fortran.
+        // Cast to size_t before multiplying to avoid int overflow for large matrices.
+        double *A_lap = new double[static_cast<size_t>(m) * n];
         double *S_lap = new double[min_mn];
-        double *U_lap = new double[m * min_mn];
-        double *VT_lap = new double[min_mn * n];
+        double *U_lap = new double[static_cast<size_t>(m) * min_mn];
+        double *VT_lap = new double[static_cast<size_t>(min_mn) * n];
 
         // Copy A into A_lap in column-major order: A_lap[i + j*m] = A(i, j)
         for (int j = 0; j < n; ++j) {
             for (int i = 0; i < m; ++i) {
-                A_lap[i + j * m] = A(i, j);
+                A_lap[static_cast<size_t>(i) + static_cast<size_t>(j) * m] = A(i, j);
             }
         }
